[IA64] Display the information in fault handler
  This patch modifies the interruption handler to display the
information about the interruption and then do panic in the case of the
fault occurred in hypervisor, or reflect to the guest when handling the
fault in guest.

  The modification is as follows:
1. In each handler in IVT, the procedure is added to branch to fault
handler or reflection according to ipsr.cpl.
2. In the case of the fault occurred in hypervisor(ipsr.cpl==0), the
fault handler displays the information (vector, registers, and stack by
calling show_registers()), and then do panic().
3. In the case of the fault occurred in guest(ipsr.cpl!=0), it is
reflected to the guest.

  With this modification, some faults which were always treated to be
reflected to the guest are changed to do panic when they occur in
hypervisor.

Signed-off-by: Tetsu Yamamoto <yamamoto.tetsu@jp.fujitsu.com>