[IA64] fix memory leak when domVTI is created
privregs clean up.
memory leak occures when VT-i domain is created.
When domain is created, xend sets max vcpu before domain setup.
So alloc_vcpu_struct() think the domain is normal domU, not domVTI.
And next xend set the domain as domVTI. so the memory is allocated for
domU won't be freed.

Signed-off-by: Isaku Yamahata <yamahata@valinux.co.jp>