[XEND] Daemonise vncviewer when it is spawned from xend.
There are various reasons it should be daemonized, but the particular
problem we hit was that YaST called "xm create" and waited on output on
stdout/stderr; xm then spawned vncviewer (which never closed its
inherited stdout and stderr); xm then would exit, but YaST still had
open file descriptors, and therefore waited forever.  It would be
possible to work around in YaST, but it seemed cleaner to daemonize
vncviewer.

Signed-off-by:  Charles Coffing <ccoffing@novell.com>