[IA64] initial cleanup of ivt.S
This patch reflects a patch I recently posted to the linux-ia64
mailing list to do essentially the same thing for ivt.S in upstream
source.

I've reformatted the contents of the Xen version of ivt.S so that
they are now readable on 80-column displays, in accordance with 
Linux coding standards.  This is really just a first pass at 
cleaning up this code.

In all of this patch, I only changed one line of assembly; there
was a 'mov r31=pr;' in entry 23 (daccess_rights) that I changed
to 'mov r31=pr;;' which is most likely what was meant.  No other
instructions were changed.  Hopefully, I didn't harm any other
white space formatting.

Signed-off-by: Al Stone <ahs3@fc.hp.com>